Linked LinkSprites to LinkAnimation and tested LinkAnimation in Game1.

+++ 3902 Project/Link/LinkSprites/ISprite.cs
@@ -9,5 +9,5 @@ public interface ISprite
 
     void Draw(SpriteBatch spriteBatch);
 
-    void Draw(SpriteBatch sb, ILinkStateMachine state, int x, int y);
+    void Draw(SpriteBatch sb, ILinkStateMachine state, double x, double y);
 }
